+++ libraries/lepton/src/ExpressionProgram.cpp
@@ -71,13 +71,13 @@ ExpressionProgram& ExpressionProgram::operator=(const ExpressionProgram& program
 }
 
 void ExpressionProgram::buildProgram(const ExpressionTreeNode& node) {
-    for (int i = node.getChildren().size()-1; i >= 0; i--)
+    for (int i = (int) node.getChildren().size()-1; i >= 0; i--)
         buildProgram(node.getChildren()[i]);
     operations.push_back(node.getOperation().clone());
 }
 
 int ExpressionProgram::getNumOperations() const {
-    return operations.size();
+    return (int) operations.size();
 }
 
 const Operation& ExpressionProgram::getOperation(int index) const {
